+++ CHANGELOG.md
@@ -5,6 +5,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Implementation for AIS VDM/VDO sentence type 10 parsing
 - Implementation for AIS VDM/VDO sentence type 9 parsing
 ### Changed
 - Dependency `env_logger` upgraded to version 0.8